requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
requests==2.31.0
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
jinja2==3.1.2
python-docx==1.1.0
pyarrow==14.0.2
openpyxl==3.1.2
